requirements.txt
flask
PyMuPDF
google-generativeai
python-dotenv
